requirements.txt
pytest
azure-mgmt-resource
azure-identity